+++ src/features/TabIcons.ts
@@ -72,8 +72,7 @@ export class TabIcons {
 		const img = document.createElement("img");
 		img.width = TAB_ICON_SIZE;
 		img.height = TAB_ICON_SIZE;
-		const adapter = this.plugin.app.vault.adapter;
-		img.src = adapter.getResourcePath(`${this.plugin.manifest.dir}/icons/${icon.value}.png`);
+		img.src = this.plugin.iconLibrary.getIconUrl(icon.value);
 		img.alt = "";
 		wrapper.appendChild(img);